updated Transaction

+++ src/main/java/org/Backend/Entity/Transaction.java
@@ -24,7 +24,7 @@ public class Transaction {
         return type;
     }
 
-    public Long getTransactionNumber() {
+    public int getTransactionNumber() {
         return transactionNumber;
     }
